Assault Update
Much nicer.
A state change function would be nice to be called in chooseAction. It
should supply the correct state to update to for the FSM which will
then dictate the action function to call.
In the case of assault, the mode should terminated as soon as it
crosses the line. this line can be calculated using the method in
beginAssault, I can handle it after the next state is implemented if
needle.
Cheers